safari-600.7-branch/LayoutTests/media/media-source/media-source-closed-expected.txt


EXPECTED (source.readyState == 'closed') OK
EVENT(sourceopen)
RUN(video.src = "")
EVENT(sourceclose)
END OF TEST
